Fix screensaver dbus name not found in MATE desktop
The dbus interface name under MATE desktop seems to be
`org.mate.ScreenSaver` instead of `org.freedesktop.ScreenSaver`.
This caused an uncaught dbus.exceptions.DBusException exception to
bubble up to main and crash, as reported in Github issue #36.

Fixed by refactoring the search for dbus interface to its own function
and adding the MATE desktop path to a list of DBus path to check

Note that the exception still bubbles up if still not found