Fix unnecessary audio transcoding
When I tried to prepare a video so that no transcoding was necessary, I noticed
this was impossible, even if I used exactly the same `ffmpeg` call gnomecast was
using.  I chased this down to an `or` that probably should have been an `and`:
we want to transcode if we have an unsupported audio format *and* we can't play
the embedded audio stream.